separate Theme Listeners into separate classes for easier config and testability (later on)
